+++ logs.js
@@ -1,4 +1,4 @@
- const logs = {
+ export const logs = {
     start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
     end: [
         'Результат удара [playerWins]: [playerLose] - труп',
@@ -37,4 +37,8 @@
     ],
     draw: 'Ничья - это тоже победа!'
 };
-export default logs;
+export const HIT = {
+    head: 30,
+    body: 25,
+    foot: 20,
+}
