+++ style.css
@@ -408,3 +408,15 @@ input[type=radio]:checked + label > span::before {
 input[type=radio]:checked + label > span:hover {
     background: none;
 }
+.winTitle{
+    position: absolute;
+    margin-top: 130px;
+   left: 47%;
+    font-family: 'Mortal Kombat 3';
+    font-size: 26px;
+    line-height: 20px;
+    letter-spacing: 2px;
+    color: #efefef;
+    top: 0;
+    text-transform: uppercase;
+}
